Re-order uploader filter
Dedup: transform filename into SHA-256 of the file
Exiftool: strip GPS information from EXIF

Previously Exiftool was applied after Dedup, making the filename not
match its hash (it's the hash of the photo with intact EXIF). There is
also a negligible possibility that GPS information in the EXIF can be
restored by enumerating all possible GPS coordiates and compare the
hash.

Now Exiftool is applied as the first step.